To many things that i update and modified, its for MVP phase 2

+++ app/Http/Controllers/AnakController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Anak;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class AnakController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
         $anaks = Anak::all();
@@ -15,11 +21,18 @@ class AnakController extends Controller
 
     public function create()
     {
+        if (Auth::user()->role != 'admin') {
+            return redirect('/anak')->with('error', 'Unauthorized Access');
+        }
         return view('anak.create');
     }
 
     public function store(Request $request)
     {
+        if (Auth::user()->role != 'admin') {
+            return redirect('/anak')->with('error', 'Unauthorized Access');
+        }
+
         $request->validate([
             'nama_lengkap' => 'required',
             'jenis_kelamin' => 'required',
@@ -40,11 +53,18 @@ class AnakController extends Controller
 
     public function edit(Anak $anak)
     {
+        if (Auth::user()->role != 'admin') {
+            return redirect('/anak')->with('error', 'Unauthorized Access');
+        }
         return view('anak.edit', compact('anak'));
     }
 
     public function update(Request $request, Anak $anak)
     {
+        if (Auth::user()->role != 'admin') {
+            return redirect('/anak')->with('error', 'Unauthorized Access');
+        }
+
         $request->validate([
             'nama_lengkap' => 'required',
             'jenis_kelamin' => 'required',
@@ -60,8 +80,11 @@ class AnakController extends Controller
 
     public function destroy(Anak $anak)
     {
+        if (Auth::user()->role != 'admin') {
+            return redirect('/anak')->with('error', 'Unauthorized Access');
+        }
+
         $anak->delete();
         return redirect()->route('anak.index');
     }
 }
-
